requirements.txt
python-dotenv
ccxt
python-telegram-bot
pandas
numpy
schedule
